Close #56: Implemented accounting unit type feature Landing tax rule consider flight time now

+++ src/FLS.Server.Service/Accounting/RuleEngines/DeliveryItemRulesEngine.cs
@@ -56,7 +56,7 @@ namespace FLS.Server.Service.Accounting.RuleEngines
                 rules.Add(rule);
             }
 
-            _ruleBasedDelivery.ActiveFlightTimeInSeconds = Convert.ToInt64(_flight.FlightDurationZeroBased.TotalSeconds);
+            _ruleBasedDelivery.ActiveFlightTimeInSeconds = Convert.ToInt32(_flight.FlightDurationZeroBased.TotalSeconds);
 
             while (_ruleBasedDelivery.ActiveFlightTimeInSeconds > 0)
             {
